# frp-backend/internal/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ashan FRP API",
    "version": "2.0.0",
    "description": "Go-first FRP management plane for Unraid NAS. Single binary, embedded UI, API-first control surface.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token returned by POST /auth/login. Send as 'Authorization: Bearer <token>'."
      }
    },
    "schemas": {
      "APIError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {}
        }
      },
      "ResponseEnvelope": {
        "type": "object",
        "properties": {
          "data": {},
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              },
              "trace_id": {
                "type": "string"
              },
              "job": {
                "$ref": "#/components/schemas/JobSummary"
              }
            }
          },
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      },
      "JobSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "account": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "login_name": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          },
          "auth": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string"
              },
              "mode": {
                "type": "string"
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "PasswordChangeRequest": {
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "properties": {
          "old_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "CAARecordData": {
        "type": "object",
        "required": [
          "flags",
          "tag",
          "value"
        ],
        "properties": {
          "flags": {
            "type": "integer",
            "minimum": 0,
            "maximum": 255
          },
          "tag": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "DNSRecordInput": {
        "type": "object",
        "required": [
          "type",
          "name",
          "ttl"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "A",
              "AAAA",
              "CNAME",
              "TXT",
              "MX",
              "CAA"
            ]
          },
          "name": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "description": "Required for A, AAAA, CNAME, TXT, and MX records."
          },
          "ttl": {
            "type": "integer",
            "minimum": 1,
            "maximum": 86400,
            "description": "Use 1 for automatic TTL, otherwise 60 to 86400 seconds."
          },
          "proxied": {
            "type": "boolean",
            "description": "Only valid for A, AAAA, and CNAME records."
          },
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 65535,
            "description": "Required for MX records."
          },
          "caa": {
            "$ref": "#/components/schemas/CAARecordData"
          }
        }
      },
      "DNSRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "ttl": {
            "type": "integer"
          },
          "proxied": {
            "type": "boolean"
          },
          "proxiable": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "Node": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "node_type": {
            "type": "string"
          },
          "endpoint_url": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "archived"
            ]
          },
          "health_status": {
            "type": "string",
            "enum": [
              "online",
              "degraded",
              "offline",
              "unknown"
            ]
          },
          "external_id": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NodeInput": {
        "type": "object",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "node_type": {
            "type": "string"
          },
          "endpoint_url": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "archived"
            ]
          },
          "external_id": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "Tunnel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subdomain": {
            "type": "string"
          },
          "full_domain": {
            "type": "string"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "http",
              "https",
              "tcp",
              "udp",
              "stcp",
              "xtcp"
            ]
          },
          "local_ip": {
            "type": "string"
          },
          "local_port": {
            "type": "integer"
          },
          "remote_port": {
            "type": "integer"
          },
          "desired_state": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled",
              "archived"
            ]
          },
          "actual_state": {
            "type": "string",
            "enum": [
              "pending",
              "enabled",
              "disabled",
              "error"
            ]
          },
          "state_reason": {
            "type": "string"
          },
          "chmlfrp_node": {
            "type": "string"
          },
          "chmlfrp_tunnel_name": {
            "type": "string"
          },
          "chmlfrp_tunnel_id": {
            "type": "string"
          },
          "cf_record_id": {
            "type": "string"
          },
          "onepanel_website_id": {
            "type": "integer"
          },
          "onepanel_ssl_enabled": {
            "type": "boolean"
          },
          "onepanel_proxy_target": {
            "type": "string"
          },
          "last_applied_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_error_code": {
            "type": "string"
          },
          "last_error_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TunnelInput": {
        "type": "object",
        "required": [
          "name",
          "node_id",
          "protocol"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "subdomain": {
            "type": "string"
          },
          "local_ip": {
            "type": "string"
          },
          "local_port": {
            "type": "integer"
          },
          "remote_port": {
            "type": "integer"
          },
          "chmlfrp_node": {
            "type": "string"
          },
          "desired_state": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login with username/password",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
...
    "/auth/password/change": {
      "post": {
        "summary": "Change current user password",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          }
        }
      }
    },
...
    "/auth/tokens/{id}/revoke": {
      "post": {
        "summary": "Revoke an auth token",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/nodes": {
      "get": {
        "summary": "List nodes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "nodes": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Node"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a node",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Node"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeInput"
              }
            }
          }
        }
      }
    },
...
    "/nodes/{id}": {
      "get": {
        "summary": "Get a node",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update a node",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeInput"
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/tunnels": {
      "get": {
        "summary": "List tunnels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "tunnels": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Tunnel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chmlfrp_node",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a tunnel",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TunnelInput"
              }
            }
          }
        }
      }
    },
    "/tunnels/{id}": {
      "get": {
        "summary": "Get a tunnel",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update a tunnel",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TunnelInput"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Archive a tunnel",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/tunnels/{id}/provision": {
      "post": {
        "summary": "Provision tunnel (CF DNS + 1Panel + ChmlFrp)",
        "responses": {
          "202": {
            "description": "Queued"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/settings": {
      "get": {
        "summary": "Get settings snapshot",
        "description": "Authenticated single-admin settings response. Cloudflare and ChmlFrp secrets are intentionally returned in plaintext for the personal self-hosted Settings Center; responses use Cache-Control: no-store.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "summary": "Update settings",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {}
          }
        }
      }
    },
    "/settings/integrations/cloudflare/configure": {
      "post": {
        "summary": "Auto-detect, verify, and atomically save a Cloudflare API Token or Global API Key",
        "description": "API Tokens are tested with Bearer authentication. Global API Keys use X-Auth-Email and X-Auth-Key. A single accessible zone is verified and saved automatically; multiple zones return zone_selection_required until zone_id is submitted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "secret"
                ],
                "properties": {
                  "secret": {
                    "type": "string",
                    "description": "Cloudflare API Token or Global API Key"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Required for Global API Key authentication"
                  },
                  "zone_id": {
                    "type": "string",
                    "description": "Selected zone when discovery returns more than one zone"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential saved, or a safe zone_selection_required response"
          },
          "400": {
            "description": "Credential invalid, email required, no zones, inaccessible zone, or DNS read verification failed"
          },
          "500": {
            "description": "Encrypted credential transaction failed"
          }
        }
      }
    },
    "/dns/records": {
      "get": {
        "summary": "List all Cloudflare DNS records for the configured Zone",
        "responses": {
          "200": {
            "description": "DNS records",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "zone": {
                              "type": "string"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "records": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/DNSRecord"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "412": {
            "description": "Cloudflare Token or Zone has not been configured"
          },
          "502": {
            "description": "Cloudflare DNS request failed"
          }
        }
      },
      "post": {
        "summary": "Create a Cloudflare DNS record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DNSRecordInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DNSRecord"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid DNS record input"
          },
          "412": {
            "description": "Cloudflare Token or Zone has not been configured"
          },
          "502": {
            "description": "Cloudflare DNS request failed"
          }
        }
      }
    },
    "/dns/records/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "summary": "Update a supported Cloudflare DNS record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DNSRecordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Invalid DNS record input"
          },
          "412": {
            "description": "Cloudflare Token or Zone has not been configured"
          },
          "502": {
            "description": "Cloudflare DNS request failed"
          }
        }
      },
      "delete": {
        "summary": "Delete a Cloudflare DNS record",
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "412": {
            "description": "Cloudflare Token or Zone has not been configured"
          },
          "502": {
            "description": "Cloudflare DNS request failed"
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "summary": "Jobs list",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Job detail",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/audit": {
      "get": {
        "summary": "Filter and paginate security audit logs",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "request_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  }
}